+++ NEWS.md
@@ -2,6 +2,7 @@
 # sabRmetrics 1.1.0 (in development)
 
 - Add new `download_season_summary()` function
+- Add new `download_player()` function
 
 # sabRmetrics 1.0.5 (2024-11-17)